requirements.txt
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.0
python-docx==1.0.1
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.1.0

